# OSIDB-134,OSIDB-1620,OSIDB-1621 - perf tweaks