--- a/.secrets.baseline
+++ b/.secrets.baseline
@@ -279,7 +279,7 @@
         "filename": "osidb/models.py",
         "hashed_secret": "c7e672880d394aa5dd924e04465c986652ba7291",
         "is_verified": false,
-        "line_number": 230,
+        "line_number": 231,
         "is_secret": false
       }
     ],
@@ -312,5 +312,5 @@
       }
     ]
   },
-  "generated_at": "2023-12-15T14:12:48Z"
+  "generated_at": "2024-01-10T09:34:31Z"
 }