--- a/etc/pg/postgresql.conf
+++ b/etc/pg/postgresql.conf
@@ -198,7 +198,7 @@ dynamic_shared_memory_type = posix	# the default is the first option
 #fsync = on				# flush data to disk for crash safety
 					# (turning this off can cause
 					# unrecoverable data corruption)
-#synchronous_commit = on		# synchronization level;
+synchronous_commit = off		# synchronization level;
 					# off, local, remote_write, remote_apply, or on
 #wal_sync_method = fsync		# the default is the first option
 					# supported by the operating system:
@@ -372,8 +372,8 @@ min_wal_size = 80MB
 # - Planner Cost Constants -
 
 #seq_page_cost = 1.0			# measured on an arbitrary scale
-#random_page_cost = 4.0			# same scale as above
-#cpu_tuple_cost = 0.01			# same scale as above
+random_page_cost = 1.1			# same scale as above
+cpu_tuple_cost = 0.03			# same scale as above
 #cpu_index_tuple_cost = 0.005		# same scale as above
 #cpu_operator_cost = 0.0025		# same scale as above
 #parallel_tuple_cost = 0.1		# same scale as above