* add necc prefetches to Flaw, Affect and Tracker models [[1](https://github.com/RedHatProductSecurity/osidb/pull/414/commits/5716bd10f9b0377301095e770fc1e4068b45a974)]
* add gin indexes for Row Based Security performance directly on models [[5](https://github.com/RedHatProductSecurity/osidb/pull/414/commits/350afce4b9557da4dce69cae0df133f4aa0f05c4), removed django migration (thx @Elkasitu !)]
* added MAX_CONNS=300 [[3](https://github.com/RedHatProductSecurity/osidb/pull/414/commits/4139a6d60b748d894fbf68ed09e8ca5c0a1db4bc)]
* align local pg config [[4](https://github.com/RedHatProductSecurity/osidb/pull/414/commits/3cc50b461a530db1dcf6c3f4c5e35bcad6cccd12)]

The relative change in db sql performance is (sans cache and cold query):

| endpoint         | stage before     | stage after perf-tweaks branch deployed |
|--------------|-----------|------------|
| /osidb/api/v1/flaws?limit=10  |  103 queries in 13353ms,  CPU 3420ms         |  14 queries in 2923ms,  CPU 940ms       |
| /osidb/api/v1/affects?limit=100               |  282 queries in 56601ms,  CPU 6467ms         |   7 queries in 1374ms,  CPU876ms      |
| /osidb/api/v1/trackers?limit=100             |  203 queries in 41630ms,  CPU 41630ms             |    5 queries in 1089ms,  CPU 497ms     |

Which roughly works out as being >5x faster for flaw endpoint and >10x faster for affects/trackers endpoints. All endpoints are much more efficient which will translate to better user concurrency. I expect real numbers to be faster when directly testing prod.

**Note**- elected to inject pg configs on ops repo and gin indexes as a django migration ... setting up gin indexes in Django (via model Meta definition) should be possible but in other projects we noticed it 'does not work' - hence the django migration.

**Note** - stage has very little loaded in it so real time comparisons are not going to be dramatic when tested there.