@JimFuller-RedHat 2024-01-09
@Elkasitu added perf comparison on the endpoints ... will fiddle with renaming commits once I get back all comments