@JimFuller-RedHat 2024-01-09
thx to @Elkasitu for figuring out we can set GinIndex direct on model ... made the change